Fixed edge case in King output pruning script
Was not working for edge cases where all related individuals prune down to one unpruned sample pair.
R converts that matrix into a character vector meaning we have to use different logic.
Ignoring this case caused a failure because we called nrow() on a character vector which obviously is no bueno
So now we check for this case before we check nrow().