Include response message in return statement
Previously, the code only had a bare 'return' statement, which didn't provide any information about the success or failure of the operation. This commit updates the return statement to include a JSON response body message that indicates the success of the operation and the name of the cleaned up bucket. The response body is constructed using f-strings and the 'json.dumps' method.